Bugfix use firstPartyID as the receiver of handouts Bugfix add new Handouts before removing returns to prevent issues in the condition where a handout was given and received in the same ESI fetch Bugfix adjusted WalletEntry.ID to a Long to properly fetch data Modifications on constraint layouts for displaying handouts

+++ app/src/main/java/com/example/evehandoutmanager/home/WalletEntry.kt
@@ -8,7 +8,7 @@ class WalletEntry (
     @SerializedName("date"            ) var date          : String,
     @SerializedName("description"     ) var description   : String? = null,
     @SerializedName("first_party_id"  ) var firstPartyId  : Int,
-    @SerializedName("id"              ) var id            : Int,
+    @SerializedName("id"              ) var id            : Long,
     @SerializedName("reason"          ) var reason        : String? = null,
     @SerializedName("ref_type"        ) var refType       : String,
     @SerializedName("second_party_id" ) var secondPartyId : Int,
